// src/services/api/gen/types.gen.ts
export type Profile = {
    gender?: string | null;
    aadhaar_number: string;
    dob: string;
    readonly profile_url: string | null;
};
// ...
export type UserProfile = {
    email: string;
    first_name?: string;
    last_name?: string;
    readonly phone_number: string;
    user_profile?: Profile;
};